requirements.txt
bleach==3.1.0
Django==3.1.1
django-adminrestrict==3.0
django-allauth==0.42.0
django-crispy==1.8.1
django-environ==0.4.5
django-extensions==3.0.8
django-filter==2.4.0
django-icons==2.1.1
django-mathfilters==1.0.0
django-request-logging==0.7.2
django-rest==0.8.6
django-rest-framework==0.1.0
django-search-views==0.3.7
django-storages==1.11.1
django-utils==0.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
djangorestframework-queryfields==1.0.0
environ==1.0
fuzzywuzzy==0.18.0
gunicorn==20.0.4
image==1.5.32
joblib==1.1.0
mordred==1.2.0
oauthlib==3.1.0
psycopg2-binary==2.9.3
python3-openid==3.2.0
regex==2020.2.20
requests==2.22.0
requests-oauthlib==1.3.0
runscript==0.2.9
s3transfer==0.4.2
scipy==1.7.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
TPOT==0.11.1
urllib3==1.25.10
whitenoise==5.2.0
xgboost==1.3.3
